requirements.txt
eth-account
requests
rich
loguru
python-dotenv
